NP-2020 Allow reusing emptyPagesRange
